requirements.txt
requests>=2.28.0
pymongo>=4.0.0
python-dateutil>=2.8.0
pytz>=2022.1
python-dotenv>=0.19.0
aiohttp>=3.8.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
urllib3>=1.26.0